+++ apps/client/src/features/dashboard/pages/Dashboard.tsx
@@ -20,8 +20,8 @@ import { useAppStore } from '../store/useAppStore';
 import { useIsMobile } from '@/shared/hooks/use-responsive';
 import { cn } from '@/lib/utils';
 import { Card, CardContent, CardHeader, CardTitle } from '@/shared/components/ui/card';
-import AuthStatus from '@/features/auth/components/AuthStatus';
-import { useAuthStore } from '@/features/auth/store/useAuthStore';
+import AuthStatus from '@/shared/layout/AuthStatus';
+import { useAuthStore } from '@/shared/stores/authStore';
 import { SimplifiedTaskCreation } from '../components/SimplifiedTaskCreation';
 import { EmptyState } from '@/shared/components/ui/empty-state';
 import { Button } from '@/shared/components/ui/button';
